requirements.txt
opencv-python~=4.5.5.64
matplotlib~=3.6.1
jmetalpy
